requirements_test.txt
# Requirements to test the source code
# - last checked/updated: 2025-10-27 (c.f. HA 2025.10.0)
#
# Note: requires ramses_cc v0.52.1 (pre-release) or later

# may need to use this version of ramses_rf...
# -e ~/client/ramses_rf

# for using the library with HA
# - pip list | grep -E 'pytest|voluptuous'

  -r requirements_dev.txt


# used for testing (incl. HA, pytest*, syrupy, voluptuous, etc.)
# - pytest

  pytest_homeassistant_custom_component >= 0.13.286
